requirements.txt
flask==3.0.3
flask-jwt-extended==4.6.0
flask-cors==4.0.1
passlib==1.7.4
python-dotenv==1.0.1
gunicorn==21.2.0
sendgrid==6.11.0